chore: delete other CredentialProgramConfig
